Refactor Menu Category & Product model, controller, interface & service

+++ [Spring]-RestaurantWebApp/src/main/java/argmus/restaurantwebapp/repository/MenuCategoryRepository.java
@@ -3,5 +3,5 @@ package argmus.restaurantwebapp.repository;
 import argmus.restaurantwebapp.model.MenuCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuCategoryRepository extends JpaRepository<MenuCategory, Integer> {
+public interface MenuCategoryRepository extends JpaRepository<MenuCategory, Long> {
 }
